+++ server/component/style/image/tpl_image.php
@@ -3,4 +3,4 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 ?>
-<img src="<?php echo $url; ?>" class="<?php echo $fluid; ?> <?php echo $this->css; ?>" alt="<?php echo $this->alt; ?>" title="<?php echo $this->title; ?>">
+<img src="<?php echo $url; ?>" class="<?php echo $fluid; ?> <?php echo $this->css; ?>" alt="<?php echo $this->alt; ?>" title="<?php echo $this->title; ?>" width="<?php echo $this->width; ?>", height="<?php echo $this->height; ?>">