+++ CHANGELOG.md
@@ -22,6 +22,7 @@
  - show `name` and `displayName` for tables in `admin/data`
  - #492 - do not show `user_name` in `Admin\Data`. Instead now add `user_code`
  - #474 - rename field `source` to `img_src` and make it `content` field and it can be set differently for each language
+ - #474 - add `height` and `width` fields for style `image`
 
 ### Bugfix
  - set values for `JSON` content fields properly in CMS